requirements.txt
asgiref==3.7.2
Django==4.2.8
django-cleanup==8.0.0
django-cors-headers==4.3.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
Pillow==10.1.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
whitenoise==6.6.0
